% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AGOscores}
\alias{AGOscores}
\title{Data used for k-means cluster analysis example.}
\format{
an rda file
}
\usage{
AGOscores
}
\description{
Twelve items adapted from the Achievement Goal Questionnairewere administered to n=2111 college students (see Finney, Pieper, & Barron, 2004).
}
\keyword{datasets}
