% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{animal}
\alias{animal}
\title{Rate how similar are the nine types of animals.}
\format{an rda file}
\usage{
animal
}
\description{
Someone(presumably Greg) has ranked all the possible pairs of the 9 animals in terms of their similarity. We want to know what governs this individual’s opinions about the similarity/dissimilarity between the animals.
}
\keyword{datasets}
