% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beer}
\alias{beer}
\title{Data for classifying beers.}
\format{an rda file}
\usage{
beer
}
\description{
A marketing analyst is trying to develop a classification scheme for beer based on several characteristics that may be relevant to consumers. She samples 20 popular beers and gathers data on their per-serving in terms of the following 4 aspects: calorie content, sodium content, alcohol content, and cost per unit volume.
}
\keyword{datasets}
