% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cheat_dat}
\alias{cheat_dat}
\title{Data used for the LCA example.
Dayton & Scheers (1997) surveyed n=319 upper level undergraduates
about their cheating behavior in the university setting.
"Haveyou ever lied to avoid taking an exam and/or to avoid handing in a paper on time?"
"Have you ever purchased a term paper and/or acquired a copy of an exam ahead of time?"
"Have you ever copied answers from another student during an exam?"}
\format{
an rda file
}
\usage{
cheat_dat
}
\description{
Data used for the LCA example.
Dayton & Scheers (1997) surveyed n=319 upper level undergraduates
about their cheating behavior in the university setting.
"Haveyou ever lied to avoid taking an exam and/or to avoid handing in a paper on time?"
"Have you ever purchased a term paper and/or acquired a copy of an exam ahead of time?"
"Have you ever copied answers from another student during an exam?"
}
\keyword{datasets}
