% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hotelling}
\alias{hotelling}
\title{Hotelling's data.}
\format{an rda file}
\usage{
hotelling
}
\description{
The correlation matrix for hotelling's data (1933). It is a correlation matrix from a sample of 200 seventh grade students on tests of 1) reading speed, 2) reading power, 3) arithmetic speed, and 4) arithmetic power.
}
\keyword{datasets}
