% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Hotellings_T}
\alias{Hotellings_T}
\title{Data used for two-group MANOVA (Hotelling's T^2).}
\format{an rda file}
\usage{
Hotellings_T
}
\description{
Middle school students are randomly assigned to an educational treatment or a control group. After receiving their treatment or control experience, students are evaluated in terms of cognitive and affective outcomes together. We are interested in testing whether there is any difference between the treatment/control groups regarding the set of outcomes.
}
\keyword{datasets}
