% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lab}
\alias{lab}
\title{Data used for the SVM example.}
\format{
an rda file
}
\usage{
lab
}
\description{
One binary outcome: Whether or not a patient has the heart disease.
Two predictor variables: age and cholesterol level.
}
\keyword{datasets}
