% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MANOVA2way}
\alias{MANOVA2way}
\title{Data used for two-way MANOVA.}
\format{an rda file}
\usage{
MANOVA2way
}
\description{
Researchers are interested in studying the effects of drugs and physical exercise on people’s weight loss and positive attitude. They wish to test both the main effects and interactions effects.
}
\keyword{datasets}
