% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{manova_dat}
\alias{manova_dat}
\title{Data used for three-group MANOVA.}
\format{
an rda file
}
\usage{
manova_dat
}
\description{
Middle school students are randomly assigned to one of three treatment groups. After receiving their respective treatment or control experience, students are evaluated in terms of cognitive and affective outcomes together. We are interested in testing whether there is any difference between the three treatment groups regarding the set of outcomes.
}
\keyword{datasets}
