% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MANOVA_RM}
\alias{MANOVA_RM}
\title{Data used for MANOVA with repeated measures.}
\format{
an rda file
}
\usage{
MANOVA_RM
}
\description{
Researchers are interested in studying whether the brand of sneakers affects how high individuals can jump up from the ground. A group of subjects were tested repeatedly wearing three different brands of sneakers: Reebok, Nike, and Converse.
}
\keyword{datasets}
