% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ME_dat}
\alias{ME_dat}
\title{Data used for the univariate mixture modeling example.}
\format{
an rda file
}
\usage{
ME_dat
}
\description{
Liu, Hancock, & Harring (2010) discussed how mixtures of univariate normal distributions could be used to examine potential systematic measurement problems. Head circumference measurements data were collected from children who were subject to child abuse and/or neglect.
}
\keyword{datasets}
