% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med.data}
\alias{med.data}
\title{Data for medical students.}
\format{
an rda file
}
\usage{
med.data
}
\description{
A data set containing the data from 61 fourth year medical students (Marascuilo & Levin, 1983).
}
\keyword{datasets}
