% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{teacher}
\alias{teacher}
\title{Data used for model-based MANOVA example.}
\format{
an rda file
}
\usage{
teacher
}
\description{
Testing the differences in three outcome variables (teach1-teach3) between two professor groups.
}
\keyword{datasets}
