% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Teacher_ratings}
\alias{Teacher_ratings}
\title{Data used for three-group MANOVA with six outcomes.}
\format{
an rda file
}
\usage{
Teacher_ratings
}
\description{
The faculties from a public research univeristy were evaluated at the end of a semester with regard to the courses they taught in that semester. Students rated the faculties on six aspects. We are interested in testing whether there is any difference between different types of faculties regarding the set of outcomes.
}
\keyword{datasets}
