% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test.data}
\alias{test.data}
\title{Assessment data from Hancock (1994).}
\format{an rda file}
\usage{
test.data
}
\description{
The correlation matrix reported in Hancock (1994). The examinations for two educational measurement/statistics classes were constructed as half multiple-choice (MC) and half constructed-response (CR), containing equal number of items in each format written for the knowledge (know), comprehension (comp), application (appl), and analysis (anal) levels of Bloom’s Taxonomy.
}
\references{
Hancock (1994)
}
\keyword{datasets}
