% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wav}
\alias{wav}
\title{Polytomous item response data from Finney et al. (2004).}
\format{an rda file}
\usage{
wav
}
\description{
The item response data from Finney et al. (2004). There were four scale items that are expected to measure a single factor: Working Avoidance Orientation. The responses were coded on a 3-point scale (1 = not true, 3 = very true), n=300.
}
\keyword{datasets}
